chore(Product): implement API retrieval with mock data
