+++ client/src/components/ProductCard/ProductCard.module.scss
@@ -20,7 +20,7 @@
 
     img {
       width: 100%;
-      height: 100%;
+      height: inherit;
       border-radius: $border-radius-md;
       object-fit: cover;
     }
