+++ client/src/store/index.ts
@@ -1,10 +1,14 @@
 import { configureStore } from "@reduxjs/toolkit";
 import productsReducer from "./slices/postsSlice";
+import { productsApi } from "../api/productsApi";
 
 const store = configureStore({
   reducer: {
     products: productsReducer,
+    [productsApi.reducerPath]: productsApi.reducer,
   },
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(productsApi.middleware),
 });
 
 export type RootState = ReturnType<typeof store.getState>;
